Fixed JSON Reference Error

+++ demo/src/main/java/com/example/minionlinemarket/Repository/ReviewRepo.java
@@ -1,12 +1,12 @@
 package com.example.minionlinemarket.Repository;
 
-import com.example.minionlinemarket.model.Review;
+import com.example.minionlinemarket.Model.Review;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
 
 @Repository
-public interface ReviewRepo extends JpaRepository<Review, Integer> {
+public interface ReviewRepo extends JpaRepository<Review, Long> {
     List<Review> findAllByisFlagged(boolean flagged);
 }
